requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.2
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-markdownx==3.0.1
djangorestframework==3.14.0
dropbox==11.34.0
idna==3.4
Markdown==3.4.1
Pillow==9.2.0
ply==3.11
pytz==2022.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
urllib3==1.26.12
